fix js & css ui bugs
