+++ app/controllers/concerns/decidim/action_delegator/consultations/consultations_controller_override.rb
@@ -5,7 +5,7 @@ module Decidim
     module Consultations
       module ConsultationsControllerOverride
         extend ActiveSupport::Concern
-        include ::Decidim::ActionDelegator::NeedsConsultationSnippets
+        include ::Decidim::ActionDelegator::NeedsConsultationStyles
 
         included do
           helper ::Decidim::ActionDelegator::DelegationHelper
