+++ app/overrides/decidim/consultations/consultations/_question/add_delegation_link.html.erb.deface
@@ -1,3 +1,3 @@
 <!-- insert_after "erb[loud]:contains('display_take_part_button_for(question)')" -->
 
-<%= render partial: "decidim/action_delegator/link_to_question", locals: { question: question } %>
+<%= render partial: "decidim/action_delegator/consultations/link_to_question", locals: { question: question } %>
