+++ app/overrides/decidim/consultations/questions/_vote_button/add_delegations_link.html.erb.deface
@@ -1,3 +1,3 @@
 <!-- insert_before "erb[silent]:contains('elsif question.consultation.active?')" -->
 
-<%= render partial: "decidim/action_delegator/link_to_delegations", locals: { question: question } %>
+<%= render partial: "decidim/action_delegator/consultations/questions/link_to_delegations", locals: { question: question } %>
