+++ app/overrides/decidim/consultations/questions/_vote_button/add_modal.html.erb.deface
@@ -1,3 +1,3 @@
 <!-- insert_before "erb[loud]:contains('javascript_pack_tag')" -->
 
-<%= render partial: "decidim/action_delegator/delegations_modal", locals: { question: question } %>
+<%= render partial: "decidim/action_delegator/consultations/questions/delegations_modal", locals: { question: question } %>
