+++ app/overrides/decidim/consultations/questions/_vote_modal_confirm/add_delegation_callout.html.erb.deface
@@ -1,3 +1,3 @@
 <!-- insert_after ".reveal__header" -->
 
-<%= render partial: "decidim/action_delegator/callout", locals: { consultation: question.consultation } %>
+<%= render partial: "decidim/action_delegator/consultations/questions/callout", locals: { consultation: question.consultation } %>
