+++ app/packs/entrypoints/decidim_action_delegator_admin_action_delegator.js
@@ -1,0 +1,1 @@
+import "src/decidim/action_delegator/admin/action_delegator.js";
