--- app/packs/entrypoints/decidim_action_delegator_admin_action_delegator_js.js
@@ -1,1 +1,0 @@
-import "src/decidim/action_delegator/admin/action_delegator.js.es6";
