+++ app/packs/entrypoints/decidim_action_delegator_questions.js
@@ -1,3 +1,5 @@
+import "src/decidim/action_delegator/questions.js";
+
 require.context("../images", true)
 
 import "entrypoints/decidim_action_delegator.scss";
