--- app/packs/entrypoints/decidim_action_delegator_questions_js.js
@@ -1,1 +1,0 @@
-import "src/decidim/action_delegator/questions.js.es6";
