+++ app/packs/src/decidim/action_delegator/questions.js
@@ -23,4 +23,5 @@ $(() => {
   voteButton.click(() => {
     delegationCallouts.addClass("is-hidden");
   });
+  console.log("handle vote button", voteButton);
 });
