--- app/views/decidim/action_delegator/_delegations_modal.html.erb
@@ -1,93 +1,0 @@
-<div class="reveal" data-reveal id="delegations-modal">
-  <div class="reveal__header">
-    <h3 class="reveal__title">
-      <%= t "action_delegator.delegations_modal.title", scope: "decidim" %>
-    </h3>
-    <button class="close-button" data-close aria-label="Close modal" type="button">
-      <span aria-hidden="true">&times;</span>
-    </button>
-  </div>
-
-  <p><%= t "action_delegator.delegations_modal.contextual_help", scope: "decidim" %></p>
-  <div class="card card--secondary">
-    <div class="card__content">
-
-      <% signed_in? && Decidim::ActionDelegator::GranteeDelegations.for(question.consultation, current_user).each do |delegation| %>
-        <div class="row">
-          <div class="columns medium-5 medium-offset-1">
-            <%= delegation.granter.name %>
-          </div>
-          <div class="columns medium-5 medium-offset-1">
-
-            <%# Copied from decidim-consultation's _vote_button.html.erb %>
-            <% if question.consultation.upcoming? %>
-              <div class="button expanded disabled">
-                <div class="vote-button-caption"><%= t("questions.vote_button.vote", scope: "decidim") %></div>
-                <div>
-                  <%= t "questions.vote_button.starting_from",
-                        scope: "decidim",
-                        date: l(question.start_voting_date) %>
-                </div>
-              </div>
-            <% elsif question.consultation.finished? && signed_in? && question.voted_by?(delegation.granter) %>
-              <%= button_to decidim_consultations.question_question_votes_path(question),
-                          method: :delete,
-                          remote: true,
-                          data: { disable: true },
-                          class: "button expanded success disabled delegation_unvote_button" do %>
-                <div class="vote-button-caption">
-                  <%= t("questions.vote_button.already_voted", scope: "decidim") %>
-                </div>
-              <% end %>
-            <% elsif signed_in? && question.consultation.active? %>
-              <% if allowed_to? :unvote_delegation, :question, { question: question, delegation: delegation }, [Decidim::ActionDelegator::Permissions, Decidim::Admin::Permissions, Decidim::Permissions] %>
-                <%= button_to decidim_consultations.question_question_votes_path(question),
-                              method: :delete,
-                              remote: true,
-                              data: { disable: true },
-                              class: "button expanded success delegation_unvote_button" do %>
-                              <%= hidden_field_tag "decidim_consultations_delegation_id", delegation.id, id: "decidim_consultations_delegation_id_delete_#{delegation.id}" %>
-                  <div class="vote-button-caption"
-                       data-original="<%= t("questions.vote_button.already_voted", scope: "decidim") %>"
-                       data-replace="<%= t("questions.vote_button.already_voted_hover", scope: "decidim") %>">
-                    <%= t("questions.vote_button.already_voted", scope: "decidim") %>
-                  </div>
-                <% end %>
-              <% elsif allowed_to? :vote_delegation, :question, { question: question, delegation: delegation }, [Decidim::ActionDelegator::Permissions, Decidim::Admin::Permissions, Decidim::Permissions] %>
-                <% if question.multiple? %>
-                  <%= link_to decidim_consultations.question_question_multiple_votes_path(question, delegation: delegation.id),
-                              class: "button expanded",
-                              id: "multivote_button" do %>
-                    <div class="vote-button-caption"><%= t "questions.vote_button.vote", scope: "decidim" %></div>
-                  <% end %>
-                <% else %>
-
-                  <%# This link is slightly modified from _vote_button to add stuff needed by delegations %>
-                  <%= link_to "#", class: "button expanded delegation-vote-button", data: { "delegation-id" => delegation.id, "delegation-granter-name" => delegation.granter.name } do %>
-                    <div class="vote-button-caption"><%= t "questions.vote_button.vote", scope: "decidim" %></div>
-                  <% end %>
-
-                <% end %>
-              <% else %>
-                <%= authorized_vote_modal_button(question, remote: true, class: "button expanded light button--sc") do %>
-                  <%= t("questions.vote_button.verification_required", scope: "decidim") %>
-                <% end %>
-              <% end %>
-            <% elsif question.consultation.active? %>
-              <%= logged_button_to decidim_consultations.question_question_votes_path(question),
-                                   class: "button expanded",
-                                   data: { disable: true },
-                                   id: "vote_button" do %>
-                <div class="vote-button-caption">
-                  <%= t("questions.vote_button.vote", scope: "decidim") %>
-                </div>
-              <% end %>
-            <% end %>
-
-          </div>
-        </div>
-      <% end %>
-
-    </div>
-  </div>
-</div>
