+++ app/views/decidim/action_delegator/consultations/questions/_delegations_modal.html.erb
@@ -1,0 +1,36 @@
+<div class="reveal" data-reveal id="delegations-modal">
+  <div class="reveal__header">
+    <h3 class="reveal__title">
+      <%= t "action_delegator.delegations_modal.title", scope: "decidim" %>
+    </h3>
+    <button class="close-button" data-close aria-label="Close modal" type="button">
+      <span aria-hidden="true">&times;</span>
+    </button>
+  </div>
+
+  <p><%= t "action_delegator.delegations_modal.contextual_help", scope: "decidim" %></p>
+  <div class="card card--secondary">
+    <div class="card__content">
+
+      <% signed_in? && Decidim::ActionDelegator::GranteeDelegations.for(question.consultation, current_user).each do |delegation| %>
+        <div class="row">
+          <div class="columns medium-5 medium-offset-1">
+            <%= delegation.granter.name %>
+          </div>
+          <div class="columns medium-5 medium-offset-1">
+            <%# Copied from decidim-consultation's _vote_button.html.erb %>
+            <% if question.consultation.upcoming? %>
+              <%= render "decidim/action_delegator/consultations/questions/vote_delegated_upcoming", question: question %>
+            <% elsif question.consultation.finished? && question.voted_by?(delegation.granter) %>
+              <%= render "decidim/action_delegator/consultations/questions/vote_delegated_finished", question: question %>
+            <% elsif question.consultation.active? %>
+              <%= render "decidim/action_delegator/consultations/questions/vote_delegated_active", question: question, delegation: delegation %>
+            <% end %>
+          </div>
+        </div>
+      <% end %>
+    </div>
+  </div>
+</div>
+
+<%= javascript_pack_tag "decidim_action_delegator_questions" %>
