+++ app/views/decidim/action_delegator/consultations/questions/_link_to_delegations.html.erb
