+++ app/views/decidim/action_delegator/consultations/questions/_vote_delegated_active.html.erb
@@ -1,0 +1,32 @@
+<% if allowed_to? :unvote_delegation, :question, { question: question, delegation: delegation }, [Decidim::ActionDelegator::Permissions, Decidim::Admin::Permissions, Decidim::Permissions] %>
+  <%= button_to decidim_consultations.question_question_votes_path(question),
+                method: :delete,
+                remote: true,
+                data: { disable: true },
+                class: "button expanded success delegation_unvote_button" do %>
+                <%= hidden_field_tag "decidim_consultations_delegation_id", delegation.id, id: "decidim_consultations_delegation_id_delete_#{delegation.id}" %>
+    <div class="vote-button-caption"
+         data-original="<%= t("questions.vote_button.already_voted", scope: "decidim") %>"
+         data-replace="<%= t("questions.vote_button.already_voted_hover", scope: "decidim") %>">
+      <%= t("questions.vote_button.already_voted", scope: "decidim") %>
+    </div>
+  <% end %>
+<% elsif allowed_to? :vote_delegation, :question, { question: question, delegation: delegation }, [Decidim::ActionDelegator::Permissions, Decidim::Admin::Permissions, Decidim::Permissions] %>
+  <% if question.multiple? %>
+    <%= link_to decidim_consultations.question_question_multiple_votes_path(question, delegation: delegation.id),
+                class: "button expanded",
+                id: "multivote_button" do %>
+      <div class="vote-button-caption"><%= t "questions.vote_button.vote", scope: "decidim" %></div>
+    <% end %>
+  <% else %>
+    <%# This link is slightly modified from _vote_button to add stuff needed by delegations %>
+    <%= link_to "#", class: "button expanded delegation-vote-button", data: { "delegation-id" => delegation.id, "delegation-granter-name" => delegation.granter.name } do %>
+      <div class="vote-button-caption"><%= t "questions.vote_button.vote", scope: "decidim" %></div>
+    <% end %>
+
+  <% end %>
+<% else %>
+  <%= authorized_vote_modal_button(question, remote: true, class: "button expanded light button--sc") do %>
+    <%= t("questions.vote_button.verification_required", scope: "decidim") %>
+  <% end %>
+<% end %>
