+++ app/views/decidim/action_delegator/consultations/questions/_vote_delegated_finished.html.erb
@@ -1,0 +1,9 @@
+<%= button_to decidim_consultations.question_question_votes_path(question),
+            method: :delete,
+            remote: true,
+            data: { disable: true },
+            class: "button expanded success disabled delegation_unvote_button" do %>
+  <div class="vote-button-caption">
+    <%= t("questions.vote_button.already_voted", scope: "decidim") %>
+  </div>
+<% end %>
