+++ app/views/decidim/action_delegator/consultations/questions/_vote_delegated_upcoming.html.erb
@@ -1,0 +1,8 @@
+<div class="button expanded disabled">
+  <div class="vote-button-caption"><%= t("questions.vote_button.vote", scope: "decidim") %></div>
+  <div>
+    <%= t "questions.vote_button.starting_from",
+          scope: "decidim",
+          date: l(question.start_voting_date) %>
+  </div>
+</div>
